Fix special handling for modified root component descriptor in CTT
The CTT yields the updated OCM component descrriptors (e.g. with updated
OCM repository list or image references) in the end by invoking the
`cnudie.iter.iter` function with a lookup in the target registry.
However, as this function expects the root component descriptor to be
passed-in, it won't retrieve it (again) using the lookup function.
Therefore, it is necessary that the passed-in root component descriptor
is already the modified one.

Previously, this was achieved by directly manipulating the passed-in
root component descriptor. However, doing so will cause inconsistencies
when manipulating the very same component descriptor object for multiple
different target registries.
Instead, use copies of the component descriptor and explicitly set the
updated component descriptor for the `cnudie.iter.iter` function.

Signed-off-by: Jonas Brand (8R0WNI3) <jonasbrand1712@gmail.com>