Omit padding if not required
tar spec: https://www.gnu.org/software/tar/manual/html_node/Standard.html

A tar archive contains a sequence of blocks with a fixed amount of bytes
per block (BLOCKSIZE). Thus, we add empty bytes as padding to fill the
blocks.
The expression `if (missing := tarfile.BLOCKSIZE - (uploaded_bytes % tarfile.BLOCKSIZE)):`
also evaluates to true if `uploaded_bytes` is already of BLOCKSIZE
length (or a multiple of it), ultimately creating a block of empty
bytes.

This *should* not cause problems, because the end of the tar archive is
marked with *two* blocks of empty bytes. However, manual testing
unveiled that at least the bsdtar (version 3.5.3) implementation
terminates the tar archive already after *one* block of empty bytes.

This leads to a situation where adding an unnecessary padding creates a
tar archive that, if extracted using some tar implementations, does not
create the expected files.

Signed-off-by: Philipp Heil (zkdev) <philipp.heil@sap.com>