let trusted-checkout honour "ok-to-test"-label
For security-reasons, pullrequests from untrusted sources (such as
forked repositories) are either run with limited permissions (if using
`on.pull_request` as event-source), or will by default not checkout
changes from pullrequest.

Trusted-Checkout-Action was introduced to allow acceptably secure
privileged runs for pullreqeust from trusted authors.

However, this does not work well in cases of e.g. new contributors.
Hence, implement (again, as in place for concourse) a label
`reviewed/ok-to-test`, which, if set, will have the same effect (trust
the pullrequest).

For this option to work reasonably well, event-trigger should be changed
like so:

```
on:
  pull_request_target:
    types:
      - labeled
```

This will restrict triggering to labelling-events. In order to avoid
runs being triggered from labels different than `reviewed/ok-to-test`,
jobs need to be fitted w/ a filtering-condition, in addition (this
cannot be handled on event-trigger-level, unfortunately).