add "pullrequest-trusted-helper"
For running privileged pipelines against pullrequests, we use labelling
to control which pullrequests are actually allowed to run, in order to
prevent execution of malicious pullrequests.

Add a helper-action, and -workflow, that can be used to indicate such
label needs to be set, or automatically set it (if the
pullrequest-author is trusted).