fixate trusted checkout to pullrequest-event's commit-digest
If `trusted-checkout`-action is used in conjunction w/ "label-based
trust" (where privileged pipeline-runs are triggered by adding a certain
label), there is the possibility of a race-condition (where a malicious
change could be pushed immediately after such a label is added).

Hence, always have `trusted-checkout`-action do an explicit checkout of
commit by digest as indicated by the event-source. In the majority of
cases, this will exhibit no change compared to previous behaviour.