make inclusion of subcomponents' release-notes configurable
Most repositories on github.com/gardener make use of a
"non-OCM"-mechanism to create upgrade-pullrequests, which includes a
mechanism (honoured by `release-notes`-action) to convey references to
both subcomponent-upgrade-vectors, and linking of their respective
release-notes (if any).

Hence, in those cases, adding release-notes collected using OCM-means
will lead to redundancies (there will be both references to and copies
of release-notes of (changed) sub-component-versions). To avoid this,
make collection of subcomponent-release-notes optional (controlled via
input). Default to not collecting copies of subcomponents by default to
do the "right thing" in the majority of cases (so we do not have to
rollout a lot of cfg-changes).