pass "commit-kind" to called version-callback-actions
Convey semantics of commits (to-be) created w/ version-action to
callbacks, to give them a hook to behave differently, depending on for
which commit they were called.

A concrete use-case for behaving differently between bumping / releasing
being the maintenance of a "LATEST"-file, which should be incremented
upon release, but not be bumped after release.