workaround action/checkout to handle submodules w/ ssh-url on GHE
If git-submodules are referenced using SSH-Protocol,
actions/checkout-action will not properly handle this if only token is
passed (according to documentation + code, said action does tell git to
do an url.<https-url>.insteadOf <ssh-url> - however that seems to not be
honoured for submodules).

As workaround, add a slightly hacky script that takes some assumptions
(e.g. does not support cross-github-instance-references, or submodules
not hosted on github), which overwrites each submodules's url with one
calculated from GitHub-Instance running respective GHA-Pipeline-Run.

We rely on using GitHub-Apps to create short-lived access-tokens, rather
than passing-in static tokens. As in case of submodules, different
tokens may be required (as such tokens are bound to installation, and
hence github-organisation), accept (optional) github-app-client-id and
private-key as input, and create such tokens, as needed.