allow absence of app-key
reusable workflows from cc-utils (such as `prepare.yaml`) is typically
called w/o inheriting secrets (as this is typically not needed). If
handling presence of app-id as an error if app-key is absent, this would
result in errors in such cases, requiring existing callers to
(unnecessarily) inherit secrets.

Hence, downgrade case of absent key (but present app-id) to just a
warning.