Write effective version into release notes doc during release commit
In Gardener, patch release branches usually use the release commit as
base for subsequent patch releases. As the local release notes docs are
only removed as part of the bump commit, those branches still contain
the release notes docs which have already been processed during the
release.
As moving the deletion of the files to the release commit is not easily
possible, instead write the effective version to the release notes docs
as part of the release commit. This allows the release-notes step to
only consider those release notes for the current effective version and
to ignore leftover release notes docs with a different version.

Signed-off-by: Jonas Brand (8R0WNI3) <jonasbrand1712@gmail.com>