Only consider release notes once if component is in BoM multiple times
To include the release notes of all versions of subcomponents in the
upgrade-path, the check does only include the component name now.
However, this will result in duplicate release notes in case a component
is referenced multiple times in the BoM in different versions. To only
retrieve release notes once for each component identity, keep track of
the already seen component identities.

Signed-off-by: Jonas Brand (8R0WNI3) <jonasbrand1712@gmail.com>