merge ctt-cli into ocm-package
Additional OCM-Commands are needed. `ctt` package does not seem like a
good package to host such commands, as it is itself implements a
ocm-command.

Hence, drop distribution-package for `ctt` (again), and declare
dependencies from `gardener-ocm` (`ocm`) package against ctt.

Merge CLI-parsers such as to add `replicate` and
`generate-replication-cfg` (fka simple-config) as subcommands of
`ocm-cli`.

Use separate versionfile for resulting image, such as to allow starting
with reasonable versioning. This is in sync with plan to switch to
either `ocm-lib`, or just `ocm` package on PYPI.