Honour FEDERATED_GITHUB_ACCESS_TOKEN_SERVER GitHub Actions variable
If the `FEDERATED_GITHUB_ACCESS_TOKEN_SERVER` GitHub Actions variable is
set, the GitHub OIDC Federation service will be used to create a GitHub
access token instead of using a GitHub-App directly. This requires a
respective OIDC configuration in the `oidc-federation.yaml` file of the
`.github` repository. Also, the `GARDENER_GITHUB_ACTIONS_PRIVATE_KEY`
secret must not be passed anymore but instead the `id-token: write`
permission is required (if that's not already the case anyways).

Signed-off-by: Jonas Brand (8R0WNI3) <j.brand@sap.com>