Set explicit GitHub token for `cleanup-release-branches` GitHub action
As the deletion of stale release branches might be prevented by existing
repository rulesets, one has to configure a "bypass list" accordingly.
Since the bypass list only features the configuration of roles, teams, and
apps and the default `github-actions[bot]` user is not assigned to any
role or team, we have to use an app instead. Hence, allow for passing-in
of an explicit (app-)token.

Signed-off-by: Jonas Brand (8R0WNI3) <j.brand@sap.com>