Add `workflows: write` permission to fed-token in release workflow
The `workflows: write` permission is required in case there is a concurrent head
update of the default branch affecting one of the workflow files. If the release
workflow will try to push the changes to a ref, it will be interpreted as a diff
towards the default branch and hence this permission will be required.

Signed-off-by: Jonas Brand (8R0WNI3) <j.brand@sap.com>