allow version to be omitted for --name
imagevector-command expects both a --root-name and --name parameter (both of
which specifying OCM-Component-Versions). As sub-component (specified
with `--name` must be a (transitive) subcomponent referenced from
root-component, it is acceptable for callers to omit version for the
latter (in this case, imagevector-command will default to the first
found version of the specified component-name).