ignore errors from image-sanitising
For some reason (probably related to behavioural change of docker or
buildx after recent upgrade), image-sanitise-code will run into errors
in some cases. As codepath is only run for concourse-pipelines (which
are to be dropped eventually), do not bother w/ detailed analysis, but
downgrade such errors to just a warning.